Update AddContact.py
Hi Michael, thanks for submitting the Test Task. You've got the job done in a quite straightforward manner to expected result - so well done on that. Please see some little fixes and suggested improvements for speed and readability of the code. 

Let me know if you have any questions on those and otherwise I believe we'll talk soon. 

Have a great day!